+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fix shouldNotHaveReceived with HigherOrderMessage (#842)
 * Deprecates shouldDeferMissing (#839)
 * Adds support for return type hints in Demeter chains (#848)
+* Adds shouldNotReceive to composite expectation (#847)
 
 ## 1.0.0 (2017-09-06)